Use prefix Route method to DRY the routes
The prefix() is a method of the Route facade and part of Laravel's
routing system.  The Route facade proxies calls to the router instance
(Illuminate\Routing\Router) and prefix() is a method that creates a
route group with a common URI prefix.

Inside the closure, routes are defined to automatically have the prefix
prepended.

Instead of writing /transactions/ for each route, defining it once makes
the routes more DRY (Don't Repeat Yourself) and easier to maintain.