Build new discrete routing for transactions routes
Create a new file named transactions.php in the routes directory.  This
file will contain all the transaction-related routes.

Customize the withRouting method in the bootstrap/app.php file.  This
method is part of Laravel's application bootstrapping process and
defines custom routing logic for Laravel application.  The web argument
is now an array that includes both web.php and transactions.php.  This
tells Laravel to load both files as part of the web routes.

Before and after the use of new routing, artisan route:list affirms the
transactions routes remain the same.

$ sail artisan route:list --path=transactions