Implement AssignRequestId custom middleware
Build the custom middleware by extracting X-Request-ID from the request,
or generating a new one if the request doesn't contain X-Request-ID, and
inserting it to the response.

In the global middleware context append the new custom middleware to the
global middleware.

Display X-Request-ID in the request to make sure it matches X-Request-ID
in the response.