Use built-in group in lieu of custom group
Let middleware classes (CheckUserRole, AssignRequestId, and AddColor) be
appended to the existing middleware stack of the web group.  This means
that the original middleware defined in the global web group will remain
intact, and the new middleware will be added to the end of the stack.

By appending custom middleware, we are extending the functionality of
the web group without removing or replacing its existing middleware.

The web group is typically used for routes that require session state,
CSRF protection, and other web-specific features.  It is globally
applied to most web routes in a Laravel application.  That's different
from the custom admin-access group, but the change is for demo only.