Use resolve helper in lieu of Container instance
The resolve() helper is simply a wrapper around app() designed to
improve code clarity and express intent.  Both functions resolve
dependencies from the Laravel service container.

The app() helper is a core Laravel function that provides access to the
service container.  It can be used to:

(1) Resolve a class or service from the container.
(2) Retrieve the application instance itself.

In essence, the app() helper in Laravel is a convenient way to interact
with the Laravel service container via Container class.  It provides a
shorthand for accessing the container's properties and methods without
needing to explicitly reference the container instance.

If the Container is used directly rather through the app() helper, the
same operations are carried out.