Use Stripe payment property
Build Stripe payment property and it is used to store payment-related
data (amount and transactionId) within the Stripe class.  It is
initialized as an empty array ([]) in the constructor, but it is updated
dynamically by the process method when called.  This is an organic use
of the method, better than a printout.

Without the payment property, the Stripe class would lose the ability to
retain payment data after the process method completes.