Add migration for transactions table
Use make:migration to generate transactions migration scaffolding.

$ sail artisan make:migration -- create_transactions_table

   INFO  Migration [database/migrations/2025_04_23_041722_create_transactions_table.php] created successfully.

Migrations are used to manage database schema changes.  The name of the
migration file plays a crucial role with a clear description of the
purpose of the migration.  Laravel uses the name to infer the type of
operation.  <action>_<table_name> is the migration name format.

The argument of create_transactions_table intends to create a new table
named transactions.

Now continue to build the migration by adding two lines.

(1) $table->unsignedBigInteger('user_id');

This line defines a column named user_id in the transactions table.

The unsignedBigInteger specifies that the column will store a large
unsigned integer value (64-bit).

The user_id column is intended to store the ID of a user from the users
table.  By using unsignedBigInteger, it ensures compatibility with the
id column of the users table, which is typically an auto-incrementing
unsigned big integer.

(2) $table->foreign('user_id')->references('id')->on('users');

This line establishes a foreign key constraint on the user_id column.

The foreign('user_id') specifies that the user_id column in the
transactions table is a foreign key.

The references('id') indicates that the user_id column references the id
column in another table.

The on('users') specifies the table being referenced, which is the users
table in this case.

Together the two lines establish a relationship between the transactions
table and the users table.  The user_id column in the transactions table
stores the ID of a user.  The foreign key constraint ensures that the
user_id in the transactions table always corresponds to a valid id in
the users table.