Build index of transactions
Create the index view scaffolding using make:view.

$ sail artisan make:view -- transactions/index

   INFO  View [resources/views/transactions/index.blade.php] created successfully.

Build the view with the layout generated by livewire:layout and populate
the view with the primitive PHP string.

In alignment with Laravel conventions and for compatibility with route
helpers, use plural transactions for the resource name and organize
views under resources/views/transactions/ to match the resource name.

Apply the newly built view to the index action of transaction controller
and pass the $request data to the view.