Use unescaped output
Build {!! $repo !!} to output the raw content of $repo without escaping
it.  That is, any HTML or JavaScript code within $repo will be rendered
as-is and the browser interprets the <a> tag as an actual hyperlink.

By contrast, the use of {{ $repo }} automatically escapes the content of
$repo, which dictates that any HTML or JavaScript code within $repo is
converted into plain text to prevent Cross-Site Scripting (XSS) attacks.