Build global variable
Use View::share() method to share data across all views in Laravel
application.  This makes any variable using View::share() automatically
available in every view without needing to explicitly pass it from
controllers or other parts of the application.

The key 'version' is the name of the variable that will be available in
all views.  The value '1.0.0' is the data being shared.